feat: add new abstract classes for components
BREAKING CHANGE:
nothing works after this
