+++ deeppavlov/core/models/nn_model.py
@@ -13,34 +13,28 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
-"""
-:class:`deeppavlov.models.model.Inferable` is an abstract base class that expresses the interface
-for all models that can work for inferring. The scope of all inferring models is larger than the scope
-of trainable models. For example, encoders can be inferred, but can't be trained.
-All inferring models should inherit from this class.
-"""
 from abc import abstractmethod
 
+from typing import Tuple
+
+from .component import Component
 from .serializable import Serializable
 
 
-class Inferable(Serializable):
+class NNModel(Component, Serializable):
     """
     :attr:`train_now` expresses a developer intent for whether a model as part of a pipeline
     should be trained in the current experiment run or not.
     """
 
-    def __init__(self, **kwargs):
+    def __init__(self, train_now=False, **kwargs):
+        mode = kwargs.get('mode', None)
+        if mode == 'train':
+            self.train_now = train_now
+        else:
+            self.train_now = False
         super().__init__(**kwargs)
 
     @abstractmethod
-    def infer(self, instance):
-        """
-        Infer a model. Any model can infer other model and ask it to do something (predict, encode,
-        etc. via this method)
-        :param instance: pass data instance to an inferring model
-        :param args: all needed params for inferring
-        :return a result of inferring
-        """
+    def train_on_batch(self, batch: Tuple[list, list]):
         pass
